Splitted HDF5 reader
