+++ CCDFDataModel/CCDFHDF5IO.cpp
@@ -24,6 +24,7 @@
  ******************************************************************************/
 
 #include "CCDFHDF5IO.h"
+#include "CCDFHDF5IO_ODIM.cpp"
 const char *CDFHDF5Reader::className = "CDFHDF5Reader";
 
 int CDFHDF5Reader::CustomForecastReader::readData(CDF::Variable *thisVar, size_t *start, size_t *count, ptrdiff_t *stride) {
@@ -2115,122 +2116,5 @@ int CDFHDF5Reader::convertKNMIHDF5toCF() {
   CDBDebug("convertKNMIHDF5toCF finished");
 #endif
 
-  return 0;
-}
-
-int CDFHDF5Reader::convertODIMHDF5toCF() {
-  CDBDebug("ODIM");
-
-  CDF::Variable *whereVar = cdfObject->getVariableNE("where");
-  if (whereVar == NULL) {
-    return 2;
-  }
-
-  double xScale;
-  double yScale;
-  unsigned int xSize;
-  unsigned int ySize;
-  CT::string projectionString;
-
-  CDF::Attribute *xScaleAttr = whereVar->getAttributeNE("xscale");
-  if (xScaleAttr != NULL) {
-    xScale = xScaleAttr->getDataAt<double>(0);
-  }
-  CDF::Attribute *yScaleAttr = whereVar->getAttributeNE("yscale");
-  if (yScaleAttr != NULL) {
-    yScale = yScaleAttr->getDataAt<double>(0);
-  }
-  CDF::Attribute *ySizeAttr = whereVar->getAttributeNE("ysize");
-  if (ySizeAttr != NULL) {
-    ySize = ySizeAttr->getDataAt<unsigned int>(0);
-  }
-  CDF::Attribute *xSizeAttr = whereVar->getAttributeNE("xsize");
-  if (xSizeAttr != NULL) {
-    xSize = xSizeAttr->getDataAt<unsigned int>(0);
-  }
-  CDF::Attribute *projDefAttr = whereVar->getAttributeNE("projdef");
-  if (projDefAttr != NULL) {
-    projectionString = projDefAttr->getDataAsString();
-  }
-
-  double offsetX = 0;
-  double offsetY = 0;
-
-  CDF::Variable *whatVar = cdfObject->getVariableNE("dataset1.what");
-  if (whatVar == NULL) {
-    return 2;
-  }
-
-  /* Handle time based on date and time from the HDF5 ODIM file */
-  CDF::Attribute *startDateAttr = whatVar->getAttributeNE("startdate");
-  CDF::Attribute *startTimeAttr = whatVar->getAttributeNE("starttime");
-  CT::string timeString;
-  if (startDateAttr != NULL && startTimeAttr != NULL) {
-    /* Compose the timestring based on date and time from the HDF5 ODIM file */
-    timeString.print("%sT%sZ", startDateAttr->getDataAsString().c_str(), startTimeAttr->getDataAsString().c_str());
-    CDBDebug("timeString %s", timeString.c_str());
-
-    /* Add the time dimension and timevariable */
-    CDF::Dimension *timeDim = new CDF::Dimension("time", 1);
-    cdfObject->addDimension(timeDim);
-    CDF::Dimension *varDims[] = {timeDim};
-    CDF::Variable *timeVar = new CDF::Variable(timeDim->getName().c_str(), CDF_DOUBLE, varDims, 1, true);
-    cdfObject->addVariable(timeVar);
-    timeVar->allocateData(1);
-    timeVar->setAttributeText("standard_name", "time");
-    timeVar->setAttributeText("units", "seconds since 1970-01-1");
-
-    /* Set the offset time in the time variable */
-    CTime *ctime = CTime::GetCTimeInstance(timeVar);
-    ((double *)timeVar->data)[0] = ctime->dateToOffset(ctime->freeDateStringToDate(timeString.c_str()));
-    CDBDebug("Time offset = %f", ((double *)timeVar->data)[0]);
-  }
-
-  CDBDebug("\nxScale: %f\nyScale: %f\nxSize: %d\nySize: %d\nprojdef: %s\noffsetX: %f\noffsetY: %f", xScale, yScale, xSize, ySize, projectionString.c_str(), offsetX, offsetY);
-
-  CDF::Variable *dataVar = cdfObject->getVariableNE("dataset1.data1.data");
-  if (dataVar != NULL) {
-    if (dataVar->dimensionlinks.size() == 2) {
-      dataVar->setAttributeText("grid_mapping", "crs");
-      CDF::Variable *crs = NULL;
-      crs = cdfObject->getVariableNE("crs");
-      if (crs == NULL) {
-        crs = new CDF::Variable("crs", CDF_CHAR, NULL, 0, false);
-        cdfObject->addVariable(crs);
-      }
-      crs->setAttributeText("proj4_params", projectionString.c_str());
-      CDF::Dimension *dimX = dataVar->dimensionlinks[1];
-      CDF::Dimension *dimY = dataVar->dimensionlinks[0];
-
-      if (cdfObject->getDimensionNE("time") != NULL) {
-        dataVar->dimensionlinks.insert(dataVar->dimensionlinks.begin(), cdfObject->getDimensionNE("time"));
-      }
-
-      CDF::Variable *varX = cdfObject->getVariableNE(dimX->name.c_str());
-      CDF::Variable *varY = cdfObject->getVariableNE(dimY->name.c_str());
-      varX->setName("x");
-      varY->setName("y");
-      dimX->setName("x");
-      dimY->setName("y");
-      if (CDF::allocateData(varX->currentType, &varX->data, dimX->length)) {
-        throw(__LINE__);
-      }
-      if (CDF::allocateData(varY->currentType, &varY->data, dimY->length)) {
-        throw(__LINE__);
-      }
-
-      for (size_t j = 0; j < dimX->length; j = j + 1) {
-        double x = (offsetX + double(j)) * xScale + xScale / 2 - xScale * (double(dimX->length / 2));
-        ((double *)varX->data)[j] = x;
-      }
-
-      for (size_t j = 0; j < dimY->length; j = j + 1) {
-        double y = (offsetY + float(j)) * yScale + yScale / 2 - yScale * (double(dimY->length / 2));
-        ((double *)varY->data)[j] = y;
-      }
-    } else {
-      CDBWarning("Data variable has only [%d] dimensions", dataVar->dimensionlinks.size());
-    }
-  }
   return 0;
 }
