+++ CCDFDataModel/CCDFHDF5IO.h
@@ -37,7 +37,7 @@
 #include "CProj4ToCF.h"
 #define ACCEPT_USE_OF_DEPRECATED_PROJ_API_H 1
 #include "proj_api.h"
-#define CCDFHDF5IO_DEBUG
+// #define CCDFHDF5IO_DEBUG
 
 #define CCDFHDF5IO_GROUPSEPARATOR "."
